+++ ckan/__init__.py
@@ -1,6 +1,6 @@
 # encoding: utf-8
 
-__version__ = "2.10.6"
+__version__ = "2.10.7"
 
 # The packaging system relies on this import, please do not remove it
 # type_ignore_reason: pyright thinks it's iterable