+++ ckanext/datastore/blueprint.py
@@ -9,7 +9,9 @@ from flask.wrappers import Response
 from flask.views import MethodView
 
 import ckan.lib.navl.dictization_functions as dict_fns
+from ckan.common import current_user
 from ckan.logic import (
+    check_access,
     tuplize_dict,
     parse_params,
 )
@@ -137,6 +139,12 @@ class DictionaryView(MethodView):
 
     def _prepare(self, id: str, resource_id: str) -> dict[str, Any]:
         try:
+            check_access(
+                "datastore_create",
+                context={"user": current_user.name, "auth_user_obj": current_user},
+                data_dict={"resource_id": resource_id},
+            )
+
             # resource_edit_base template uses these
             pkg_dict = get_action(u'package_show')({}, {u'id': id})
             resource = get_action(u'resource_show')({}, {u'id': resource_id})